refactor: 리뷰 반영(접근제어자 등)
