+++ src/main/java/handong/whynot/dto/common/ErrorResponseDTO.java
@@ -1,6 +1,5 @@
 package handong.whynot.dto.common;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 
 import java.util.List;
@@ -9,12 +8,11 @@ import java.util.List;
 @AllArgsConstructor
 @Getter @Setter
 @Builder
-//@JsonInclude(JsonInclude.Include.NON_NULL)   // null이면 Json반환 시 제외. (errors가 null인 경우를 위함.)
 public class ErrorResponseDTO {
 
-    Integer statusCode;
-    String message;
-    List<String> errors;
+    private Integer statusCode;
+    private String message;
+    private List<String> errors;
 
     public static ErrorResponseDTO of(ResponseCode responseCode, List<String> errors) {
         return ErrorResponseDTO.builder()
