+++ src/test/java/handong/whynot/api/PostServiceMvcTest.java
@@ -1,8 +1,9 @@
-package handong.whynot.service;
+package handong.whynot.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import handong.whynot.api.PostController;
 import handong.whynot.dto.post.PostRequestDTO;
+import handong.whynot.service.PostService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
